LayoutTests/imported/w3c/web-platform-tests/shadow-dom/execcommand-insertList-in-shadow-expected.txt


FAIL Toggle off List for all the child nodes in the ShadowRoot assert_equals: expected "Bullet One<br>Bullet Two<br>Bullet Three" but got "Bullet One<br>Bullet Two<br>Bullet Three<br>"
